Add a "managed" role if the project is managed
A bit of a hack to allow services to know if a project is managed or
not. When evaluating policies, the services (nova, at least) do not
allow passing in arbitrary information about the project. They do,
however, pull in the roles from the token.

We don't want to use real roles for managed projects, since then users
would be able to add/remove the managed role and mess things up.
Instead, if the project is managed, add a fake role to the user's roles
indicating that the project is managed.